+++ packages/iov-encoding/types/integers.d.ts
@@ -1,12 +1,18 @@
+/** Internal interface to ensure all integer types can be used equally */
 interface Integer {
   readonly toNumber: () => number;
   readonly toString: () => string;
 }
-export declare class Uint32 implements Integer {
+interface WithByteConverters {
+  readonly toBytesBigEndian: () => Uint8Array;
+  readonly toBytesLittleEndian: () => Uint8Array;
+}
+export declare class Uint32 implements Integer, WithByteConverters {
   static fromBigEndianBytes(bytes: ArrayLike<number>): Uint32;
   protected readonly data: number;
   constructor(input: number);
-  toBytesBigEndian(): readonly number[];
+  toBytesBigEndian(): Uint8Array;
+  toBytesLittleEndian(): Uint8Array;
   toNumber(): number;
   toString(): string;
 }
@@ -24,14 +30,14 @@ export declare class Uint53 implements Integer {
   toNumber(): number;
   toString(): string;
 }
-export declare class Uint64 implements Integer {
+export declare class Uint64 implements Integer, WithByteConverters {
   static fromBytesBigEndian(bytes: ArrayLike<number>): Uint64;
   static fromString(str: string): Uint64;
   static fromNumber(input: number): Uint64;
   private readonly data;
   private constructor();
-  toBytesBigEndian(): readonly number[];
-  toBytesLittleEndian(): readonly number[];
+  toBytesBigEndian(): Uint8Array;
+  toBytesLittleEndian(): Uint8Array;
   toString(): string;
   toNumber(): number;
 }